endpoints POST & GET funcionando

+++ src/main/java/com/tec/diversionesfantasy/repository/EmpleadoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.tec.diversionesfantasy.entity.Empleado;
 
 @Repository
-public interface EmpleadoRepository extends JpaRepository<Empleado, Integer>{
+public interface EmpleadoRepository extends JpaRepository<Empleado, Long>{
 
 	public Empleado findById(int id);
